requirements.txt
dlib==19.22.99
numpy==1.23.5
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
PyAutoGUI==0.9.53
tensorflow==2.12.0
tensorflow_intel==2.12.0
